+++ packages/multi-tenant/src/supertokens/recipes/third-party-email-password/index.ts
@@ -4,6 +4,7 @@ export { default as emailPasswordSignUp } from "./emailPasswordSignUp";
 export { default as emailPasswordSignUpPOST } from "./emailPasswordSignUpPost";
 export { default as generatePasswordResetTokenPOST } from "./generatePasswordResetTokenPost";
 export { default as getUserById } from "./getUserById";
+export { default as resetPasswordUsingToken } from "./resetPasswordUsingToken";
 export { default as sendEmail } from "./sendEmail";
 export { default as thirdPartySignInUp } from "./thirdPartySignInUp";
 export { default as thirdPartySignInUpPOST } from "./thirdPartySignInUpPost";