+++ packages/user/src/index.ts
@@ -33,6 +33,9 @@ export { default as userResolver } from "./model/users/resolver";
 export { default as UserService } from "./model/users/service";
 export { default as userRoutes } from "./model/users/controller";
 export { default as formatDate } from "./supertokens/utils/formatDate";
+export { default as isRoleExists } from "./supertokens/utils/isRoleExists";
+export { default as validateEmail } from "./validator/email";
+export { default as validatePassword } from "./validator/password";
 
 export type { ThirdPartyEmailPasswordRecipe } from "./supertokens/types";
 export type {